change test locations
